Refactoring the binary tree thing into a re-usable function
